Add basic ui

+++ SwiftDataExample/SwiftDataExample/Expense.swift
@@ -13,9 +13,9 @@ import SwiftData
 class Expense {
     var name: String
     var date: Date
-    var value: Int
+    var value: Double
     
-    init(name: String, date: Date, value: Int) {
+    init(name: String, date: Date, value: Double) {
         self.name = name
         self.date = date
         self.value = value
